Add rounding to numeric state variables Rename variable to _variableName Graph xlabel, ylabel are not attributes
