+++ cypress/e2e/DoenetML/tagSpecific/circle.cy.js
@@ -21981,6 +21981,56 @@ describe("Circle Tag Tests", function () {
     cy.get(cesc2("#/P3h")).should("have.text", "false");
     cy.get(cesc2("#/Q3h")).should("have.text", "true");
     cy.get(cesc2("#/R3h")).should("have.text", "false");
+  });
 
+  it("circle with rounding", () => {
+    cy.window().then(async (win) => {
+      win.postMessage(
+        {
+          doenetML: `
+    <circle center="(12.3456789, 0.123456789)" radius="1234.56789" name="c1" />
+    <p>
+      <number copySource="c1.radius" name="c1r" />
+      <number copySource="c1.diameter" name="c1d" />
+      <coords copySource="c1.center" name="c1c" />
+    </p>
+    <circle copySource="c1" displayDigits="4" displayDecimals="1" name="c1a" />
+    <p>
+      <number copySource="c1a.radius" name="c1ar" />
+      <number copySource="c1a.diameter" name="c1ad" />
+      <coords copySource="c1a.center" name="c1ac" />
+    </p>
+    <circle through="(1234.56789, 0.123456789)" name="c2" />
+    <p>
+      <point copySource="c2.throughPoint1" name="c2tp" />
+    </p>
+    <circle copySource="c2" displayDigits="4" displayDecimals="1" name="c2a" />
+    <p>
+      <point copySource="c2a.throughPoint1" name="c2atp" />
+    </p>
+    `,
+        },
+        "*",
+      );
+    });
+
+    cy.get(cesc2("#/c1r")).should("have.text", "1234.57");
+    cy.get(cesc2("#/c1d")).should("have.text", "2469.14");
+    cy.get(cesc2("#/c1c") + " .mjx-mrow")
+      .eq(0)
+      .should("have.text", "(12.35,0.123)");
+
+    cy.get(cesc2("#/c1ar")).should("have.text", "1234.6");
+    cy.get(cesc2("#/c1ad")).should("have.text", "2469.1");
+    cy.get(cesc2("#/c1ac") + " .mjx-mrow")
+      .eq(0)
+      .should("have.text", "(12.35,0.1235)");
+
+    cy.get(cesc2("#/c2tp") + " .mjx-mrow")
+      .eq(0)
+      .should("have.text", "(1234.57,0.123)");
+    cy.get(cesc2("#/c2atp") + " .mjx-mrow")
+      .eq(0)
+      .should("have.text", "(1234.6,0.1235)");
   });
 });
