+++ src/Core/components/abstract/Input.js
@@ -80,10 +80,6 @@ export default class Input extends InlineComponent {
 
     stateVariableDefinitions.creditAchieved = {
       defaultValue: 0,
-      public: true,
-      shadowingInstructions: {
-        createComponentOfType: "number",
-      },
       forRenderer: true,
       returnDependencies: () => ({
         answerAncestor: {
