+++ src/Media/bafkreibhvb4a7olqqdfk2i6dsvitlneojzbgskiwtiigxo6j3p2kzoacha.doenet
@@ -14,8 +14,8 @@
     <customAttribute componentType="number" attribute="ymax" defaultValue="10" assignNames="ymax" />
     <customAttribute componentType="boolean" attribute="showNavigation" defaultValue="true" assignNames="showNavigation" />
     <customAttribute componentType="number" attribute="attractThreshold" defaultValue="0.5" assignNames="attractThreshold" />
-    <customAttribute componentType="variable" attribute="variable" defaultValue="x" assignNames="variable" />
-    <customAttribute componentType="variable" attribute="timeVariable" defaultValue="n" assignNames="timeVariable" />
+    <customAttribute componentType="math" attribute="variable" defaultValue="x" assignNames="variable" />
+    <customAttribute componentType="math" attribute="timeVariable" defaultValue="n" assignNames="timeVariable" />
     <customAttribute componentType="boolean" attribute="checkInitialCondition" defaultValue="true" assignNames="checkInitialCondition" />
     <customAttribute componentType="number" attribute="initialConditionWeight" defaultValue="1" assignNames="initialConditionWeight" />
     <customAttribute componentType="number" attribute="cobwebbingWeight" defaultValue="1" assignNames="cobwebbingWeight" />
@@ -24,7 +24,7 @@
     <customAttribute componentType="point" attribute="defaultPoint" defaultValue="" assignNames="defaultPoint" />
   </setup>
 
-  <copy uri="doenet:cid=bafkreifax453byfe4snhlzcmxpcisfl2yfijcgsrfdxen522bn6aigap24&doenetId=2dnQm1B4zsbGzreEn0GKa" assignNames="cobwebApplet" function="$f" x0="$x0" initialInitialValue="$initialInitialValue" initialValueDx="$initialValueDx" initialValueOffset="$initialValueOffset" nIterationsRequired="$nIterationsRequired" width="$width" height="$height" xmin="$xmin" xmax="$xmax" ymin="$ymin" ymax="$ymax" showNavigation="$showNavigation" attractThreshold="$attractThreshold" variable="$variable" timeVariable="$timeVariable" showCalculatedValues="$showCalculatedValues" grid="$grid" defaultPoint="$defaultPoint" />
+  <copy uri="doenet:cid=bafkreibrckac6yd2lvw6zu2dm4ljy47vjtgqjh5tpnom3nwfsihzc2xpp4&doenetId=2dnQm1B4zsbGzreEn0GKa" assignNames="cobwebApplet" function="$f" x0="$x0" initialInitialValue="$initialInitialValue" initialValueDx="$initialValueDx" initialValueOffset="$initialValueOffset" nIterationsRequired="$nIterationsRequired" width="$width" height="$height" xmin="$xmin" xmax="$xmax" ymin="$ymin" ymax="$ymax" showNavigation="$showNavigation" attractThreshold="$attractThreshold" variable="$variable" timeVariable="$timeVariable" showCalculatedValues="$showCalculatedValues" grid="$grid" defaultPoint="$defaultPoint" />
 
   <conditionalContent condition="$checkInitialCondition">
     <p>Initial condition:
